requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
typer[all]>=0.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
textual>=0.47.0
slither-analyzer>=0.10.0
duckduckgo-search>=4.0.0
PySocks>=1.7.1
cryptography>=41.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
ruff>=0.1.0
black>=23.0.0
mypy>=1.7.0
bandit>=1.7.0
networkx
